Refactor Docker configuration forms and centralize SSH elements
Remove DockerConnectionForm and replace it with reusable components DockerWatchCard and DockerAdvancedConnectionCard. This update also renames SSHConnectionForm to SSHConnectionFormElements in related modal files, enhancing code modularity and reuse.
